first steps towards to using llvm 4.0
