+++ lib/bap_llvm/llvm_macho_loader.hpp
@@ -355,7 +355,8 @@ void dynamic_relocations(const macho &obj, command_info &info, ogre_doc &s) {
     }
 }
 
-#if LLVM_VERSION_MAJOR == 3 && LLVM_VERSION_MINOR == 8
+#if LLVM_VERSION_MAJOR == 3 && LLVM_VERSION_MINOR == 8 \
+    || LLVM_VERSION_MAJOR == 4 && LLVM_VERSION_MINOR == 0
 
 commands macho_commands(const macho &obj) {
     commands cmds;
